third_party/WebKit/LayoutTests/fast/mediastream/MediaStreamTrack-observer-iterate-no-crash.html

<!DOCTYPE HTML>
<title>Verify that MediaStreamTracks created while dispatching MediaStreamSource
 events do not crash.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
  async_test(test => {
    navigator.mediaDevices.getUserMedia({audio:true, video: true})
      .then(test.step_func(stream => {
        var track = stream.getTracks()[0];
        assert_equals(track.readyState, "live");
        track.onended = test.unreached_func("ended event should not be fired");
        track.onmute = test.step_func(() => {
          assert_equals(track.readyState, "live");
          track.clone();
          track.stop();
          assert_equals(track.readyState, "ended");
          track.clone(undefined);
  
          // Introduce a small timeout to detect a misfire of the ended event.
          setTimeout(test.step_func_done(() => {
            assert_equals(track.readyState, "ended");
          }, 100));
        });
        track.enabled = false;
      }))
      .catch(test.unreached_func("getUserMedia failed"));
  });
</script>
